+++ raiden-dapp/src/mixins/navigation-mixin.ts
@@ -93,6 +93,12 @@ export default class NavigationMixin extends Vue {
     });
   }
 
+  navigateToWithdrawal() {
+    this.$router.push({
+      name: RouteNames.ACCOUNT_WITHDRAWAL
+    });
+  }
+
   onGeneralBackClicked() {
     this.$router.go(-1);
   }