+++ raiden-ts/src/config.ts
@@ -1,9 +1,10 @@
 import * as t from 'io-ts';
-import { Network } from 'ethers/utils';
+import { Network, parseEther } from 'ethers/utils';
 
 import { Capabilities } from './constants';
 import { Address, UInt } from './utils/types';
 import { getNetworkName } from './utils/ethers';
+import { Caps } from './transport/types';
 
 const RTCIceServer = t.type({ urls: t.union([t.string, t.array(t.string)]) });
 
@@ -33,7 +34,9 @@ const RTCIceServer = t.type({ urls: t.union([t.string, t.array(t.string)]) });
  * - confirmationBlocks - How many blocks to wait before considering a transaction as confirmed
  * - monitoringReward - Reward to be paid to MS, in SVT/RDN; use Zero or null to disable
  * - logger - String specifying the console log level of redux-logger. Use '' to silence.
- * - caps - Own transport capabilities
+ * - caps - Own transport capabilities overrides. Set to null to disable all, including defaults
+ * - fallbackIceServers - STUN servers to be used as a fallback for WebRTC
+ * - rateToSvt - Exchange rate between tokens and SVT, in wei: e.g. rate[TKN]=2e18 => 1TKN = 2SVT
  * - matrixServer? - Specify a matrix server to use.
  * - subkey? - When using subkey, this sets the behavior when { subkey } option isn't explicitly
  *             set in on-chain method calls. false (default) = use main key; true = use subkey
@@ -61,8 +64,9 @@ export const RaidenConfig = t.readonly(
         warn: null,
         error: null,
       }),
-      caps: t.readonly(t.record(t.string /* Capabilities */, t.any)),
+      caps: t.union([t.null, Caps]),
       fallbackIceServers: t.array(RTCIceServer),
+      rateToSvt: t.record(t.string, UInt(32)),
     }),
     t.partial({
       matrixServer: t.string,
@@ -102,15 +106,16 @@ export function makeDefaultConfig(
     matrixExcessRooms: 3,
     pfsSafetyMargin: 1.0,
     confirmationBlocks: 5,
-    monitoringReward: null,
+    // SVT also uses 18 decimals, like Ether, so parseEther works
+    monitoringReward: parseEther('5') as UInt<32>,
     logger: 'info',
     caps: {
       [Capabilities.NO_DELIVERY]: true,
-      [Capabilities.NO_RECEIVE]: true,
       [Capabilities.NO_MEDIATE]: true,
       [Capabilities.WEBRTC]: true,
     },
     fallbackIceServers: [{ urls: 'stun:stun.l.google.com:19302' }],
+    rateToSvt: {},
     ...overwrites,
   };
 }